Spruce up tictactoe
This embodies a few changes. The player arg is removed, because I
already know which player's turn it is. I only include the actions
property if the game is not completed. I use a type of select for the
move action, and use objects instead of strings to represent the
objects. This is in line with the idea posted on an issue at the siren
github repository.